backend/requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.2.0
amqp==5.1.1
asgiref==3.7.2
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
Automat==20.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
boto3==1.24.33
botocore==1.27.33
cachetools==5.2.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cron-descriptor==1.4.0
cryptography==37.0.2
cssselect==1.1.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.2.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.0.0
django-extensions==3.2.1
django-storages==1.13.2
django-timezone-field==5.0
djangorestframework==3.14.0
exceptiongroup==1.0.0rc9
Faker==14.0.0
filelock==3.8.0
frozenlist==1.3.0
GDAL==3.4.1
geoip2==4.6.0
google-api-core==2.8.2
google-api-python-client==2.84.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-cloud-recaptcha-enterprise==1.8.0
googleapis-common-protos==1.56.4
greenlet==2.0.1
grpcio==1.47.0
grpcio-status==1.47.0
gunicorn==20.1.0
h11==0.14.0
html-sanitizer==1.9.3
httplib2==0.22.0
hyperlink==21.0.0
idna==3.3
incremental==22.10.0
iniconfig==2.0.0
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==1.0.1
kombu==5.2.4
lxml==4.9.1
maxminddb==2.2.0
multidict==6.0.2
mysqlclient==2.1.0
names==0.3.0
numpy==1.23.2
oauth2client==4.1.3
oauthlib==3.2.0
openai==0.26.0
outcome==1.2.0
packaging==21.3
parsel==1.6.0
Pillow==9.2.0
playwright==1.30.0
pluggy==1.0.0
prompt-toolkit==3.0.31
Protego==0.2.1
proto-plus==1.20.6
protobuf==3.20.1
py-spy==0.3.14
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
pyee==9.0.4
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.2.1
pytest-base-url==2.0.0
pytest-playwright==0.3.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-environ==0.4.54
python-http-client==3.3.7
python-magic==0.4.27
python-slugify==7.0.0
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
queuelib==1.6.2
requests==2.28.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
scrapy-playwright==0.0.25
selenium==4.8.0
sendgrid==6.9.7
service-identity==21.1.0
six==1.16.0
slack-sdk==3.21.2
sniffio==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
stripe==4.1.0
text-unidecode==1.3
tinycss2==1.1.1
tldextract==3.4.0
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
twilio==7.16.0
Twisted==22.8.0
typing_extensions==4.4.0
tzdata==2022.7
ua-parser==0.15.0
uritemplate==4.1.1
urllib3==1.26.9
user-agents==2.2.0
vine==5.0.0
w3lib==2.0.1
wcwidth==0.2.5
webdriver-manager==3.8.5
webencodings==0.5.1
Werkzeug==2.1.2
wsproto==1.2.0
yarl==1.7.2
zope.interface==5.5.0
